refactor: move some prof tests to better place
